+++ modules/default.nix
@@ -1,6 +1,7 @@
-{ pkgs }:
+{ pkgs, lib }:
 {
-  extraContents = import ./cdda-extras.nix { inherit pkgs; };
+  cddaExtraContents = import ./cdda-extras.nix { inherit pkgs lib; };
   cddaGameSettings = import ./cdda-game-settings.nix;
-  devshell = import ./devel/devshell.nix { inherit pkgs; };
+  devshell = import ./devel/devshell.nix { inherit pkgs lib; };
+  contentsInstaller = import ./contents-installer.nix { inherit pkgs lib; };
 }