+++ modules/cdda-extras.nix
@@ -60,25 +60,9 @@
     }
   ];
 
-  soundPacks = [
+  soundPacks = [ ];
 
-    # Example of adding Otopack's sound pack.
-    /*
-      {
-        name = "Otopack";
-        subdirs = [ "Otopack+ModsUpdates" ];
-        src = pkgs.fetchzip {
-          url = "https://github.com/Kenan2000/Otopack-Mods-Updates/"
-            + "archive/refs/tags/"
-            + "Otopack+ModsUpdates_09.03.2024.tar.gz";
-          hash =
-            "sha256-CzqDyPsFWKb6gJYserVd2X8nfJY2cugQNfC/0opLdvo=";
-        };
-      }
-    */
-  ];
-
-  # also known as gfx
+  # also known as gfx(s)
   tileSets = [
     {
       name = "UndeadPeople";